requirements.txt
pyautogui==0.9.54
Pillow==10.2.0
python-dateutil==2.8.2
opencv-python==4.9.0.80
